requirements.txt
youtube_dl==2021.12.17
gunicorn==20.1.0
Flask==2.0.3
requests==2.27.1
pandas==1.4.1
joblib==1.1.0
numpy==1.22.2
scipy==1.8.0
scikit-learn==1.0.2
lightgbm==3.3.2
python-dotenv==0.19.2
db-sqlite3==0.0.1